parse events with ABI and display the results
