+++ app/ts/types/eip721.ts
@@ -1,7 +1,7 @@
 import * as funtypes from 'funtypes'
 import { JSONEncodeableObject, isJSON } from '../utils/json.js'
-import { EthereumData, EthereumQuantity, serialize } from './wire-types.js'
-import { AddressBookEntry } from './addressBookTypes.js'
+import { EnrichedGroupedSolidityType } from './solidityType.js'
+import { serialize } from './wire-types.js'
 
 export type EIP712MessageUnderlying = funtypes.Static<typeof EIP712MessageUnderlying>
 export const EIP712MessageUnderlying = funtypes.ReadonlyObject({
@@ -30,27 +30,10 @@ const EIP712MessageParser: funtypes.ParsedValue<funtypes.String, EIP712MessageUn
 export type EIP712Message = funtypes.Static<typeof EIP712Message>
 export const EIP712Message = funtypes.String.withParser(EIP712MessageParser)
 
-export type GroupedSolidityType = funtypes.Static<typeof GroupedSolidityType>
-export const GroupedSolidityType = funtypes.Union(
-	funtypes.ReadonlyObject({ type: funtypes.Literal('integer'), value: EthereumQuantity }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('bytes'), value: EthereumData }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('fixedBytes'), value: EthereumData }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('bool'), value: funtypes.Boolean }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('string'), value: funtypes.String }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('address'), value: AddressBookEntry }),
-	
-	funtypes.ReadonlyObject({ type: funtypes.Literal('integer[]'), value: funtypes.ReadonlyArray(EthereumQuantity) }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('bytes[]'), value: funtypes.ReadonlyArray(EthereumData) }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('fixedBytes[]'), value: funtypes.ReadonlyArray(EthereumData) }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('bool[]'), value: funtypes.ReadonlyArray(funtypes.Boolean) }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('string[]'), value: funtypes.ReadonlyArray(funtypes.String) }),
-	funtypes.ReadonlyObject({ type: funtypes.Literal('address[]'), value: funtypes.ReadonlyArray(AddressBookEntry) }),
-)
-
-type typeEnrichedEIP712MessageRecord = GroupedSolidityType | { type: 'record', value: { [x: string]: typeEnrichedEIP712MessageRecord | undefined } } | { type: 'record[]', value: ReadonlyArray<{ [x: string]: typeEnrichedEIP712MessageRecord | undefined }> }
+type typeEnrichedEIP712MessageRecord = EnrichedGroupedSolidityType | { type: 'record', value: { [x: string]: typeEnrichedEIP712MessageRecord | undefined } } | { type: 'record[]', value: ReadonlyArray<{ [x: string]: typeEnrichedEIP712MessageRecord | undefined }> }
 export type EnrichedEIP712MessageRecord = funtypes.Static<typeof EnrichedEIP712MessageRecord>
 export const EnrichedEIP712MessageRecord: funtypes.Runtype<typeEnrichedEIP712MessageRecord> = funtypes.Lazy(() => funtypes.Union(
-	GroupedSolidityType,
+	EnrichedGroupedSolidityType,
 	funtypes.ReadonlyObject({ type: funtypes.Literal('record'), value: funtypes.ReadonlyRecord(funtypes.String, EnrichedEIP712MessageRecord) }),
 	funtypes.ReadonlyObject({ type: funtypes.Literal('record[]'), value: funtypes.ReadonlyArray(funtypes.ReadonlyRecord(funtypes.String, EnrichedEIP712MessageRecord)) }),
 ))
